...updating JSON structure...
Now that I look at that TreeNode file (wasnt mentioned in the
instructions BTW). I see that if I make the JSON keys match what is
already in the TreeNode file I can write less (and cleaner) code.

Normally I would never tightly couple the data to the display logic
which is why I am going to leave the HTML with the classnames I started
with.