Thinking out loude with code, Interactions and persitance
So I started off thinking this would be easy.

I am typing here so can see where my mind is. I was thinking I could use
some system like node,level,item to find a node in the tree, and insert
or remove a node. This however is not a path i can just access.

Trying to do a bunch of array operations also seems complicated.

My gut tells me that maybe as we get the dat aI could create a hashmap
with the keys being a string `nodelevelitem` and the value being the
leaf of that node.

Adding would involve finding the parent node appending to the child and
adding an entry to hashmap.

Removing would involve finding the item in the hashmap, recurisvly going
through all the children removing them and updating the hashmap
accordingly.

So far my mind/gut have not agreed on where to go. Normally I would
consult with a fellow engineer a rubber duck / whiteboard.

I am going to ponder this but I know I am way past the one hour time
frame and I still need to add dots and a few other things.

PERSITANCE
If I do process the JSON data into a hashmap, I may need to create a
function the reconsistents a new JSON object from the hashmap key and
data...

.. all of this is goign into my thoughts